Hotfix: 비밀번호 변경 화면 버그 수정 (Sangjun Lee) 2024-07-02
비밀번호 변경 화면에서 로그인 화면으로 이동 링크를 클릭 시 뒤로가기를 통해 다시 비밀번호 변경 페이지로 이동 가능하던 버그를
수정. 이제 로그인 링크를 탈 경우 모든 세션이 삭제되어 다시 비밀번호 변경 화면으로 돌아갈 수 없음.
