+++ planttycoon/src/main/resources/application-local.properties
@@ -1,0 +1,4 @@
+smtp.host=smtp.naver.com
+smtp.port=465
+smtp.username=blackyokel@naver.com
+smtp.password=72GYPBEDWU3M
