+++ planttycoon/src/main/webapp/WEB-INF/spring/root-context.xml
@@ -47,7 +47,7 @@
 	
 	<!-- Property Placeholder Configuration -->
     <bean id="propertyConfigurer" class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
-        <property name="location" value="classpath:mail-config.properties" />
+        <property name="location" value="classpath:application-local.properties" />
         <property name="fileEncoding" value="UTF-8" />
     </bean>
