+++ planttycoon/src/main/webapp/WEB-INF/web.xml
@@ -8,7 +8,7 @@
 		<param-name>contextConfigLocation</param-name>
 		<param-value>
 			/WEB-INF/spring/root-context.xml
-			/WEB-INF/spring/security-context.xml	
+			/WEB-INF/spring/security-context.xml
 		</param-value>
 	</context-param>
