+++ planttycoon/src/test/resources/root-context-test.xml
@@ -46,16 +46,17 @@
 	</bean>
     
     <!-- Property Placeholder Configuration -->
-    <!-- <bean id="propertyConfigurer" class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
-        <property name="location" value="classpath:application-ci.properties" />
+    <bean id="propertyConfigurer" class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
+        <property name="location" value="classpath:application-local.properties" />
         <property name="ignoreUnresolvablePlaceholders" value="true" />
-    </bean> -->
+    </bean>
     
+    <!-- Mail sender configuration -->
     <bean id="mailSender" class="org.springframework.mail.javamail.JavaMailSenderImpl">
-		<property name="host" value="${SMTP_HOST}"/>
-        <property name="port" value="${SMTP_PORT}"/>
-        <property name="username" value="${SMTP_USERNAME}"/>
-        <property name="password" value="${SMTP_PASSWORD}"/>
+		<property name="host" value="${smtp.host}"/>
+        <property name="port" value="${smtp.port}"/>
+        <property name="username" value="${smtp.username}"/>
+        <property name="password" value="${smtp.password}"/>
 		<property name="defaultEncoding" value="utf-8" />
 
 		<property name="javaMailProperties">
